src/pets/pets.controller.spec.ts
import { Test } from '@nestjs/testing';
import { PetsController } from './pets.controller';
import { PetsService } from './pets.service';
import { Pet } from './pets.entity';
import { getRepositoryToken } from '@nestjs/typeorm';
import { User } from '../users/users.entity';
import { CreatePetDTO } from './dtos/create-pet.dto';

describe('PetsController', () => {
  let petsController: PetsController;
  let petsService: PetsService;

  const mockRepository = {
    findOne: jest.fn(),
  };

  beforeEach(async () => {
    const module = await Test.createTestingModule({
      controllers: [PetsController],
      providers: [
        {
          provide: getRepositoryToken(User),
          useValue: mockRepository,
        },
        {
          provide: getRepositoryToken(Pet),
          useValue: mockRepository,
        },
        PetsService,
      ],
    }).compile();
    petsController = module.get<PetsController>(PetsController);
    petsService = module.get<PetsService>(PetsService);
  });

  afterEach(() => {
    jest.resetAllMocks();
  });

  describe('fetchPets with valid id', () => {
    it('should return an array of pets', async () => {
      const result: Promise<Partial<Pet>[]> = new Promise(resolve =>
        resolve([{ id: 1, name: 'Spot', type: 'DOG' }]),
      );
      jest.spyOn(petsService, 'getAll').mockImplementation(() => result);
      expect(await petsController.fetchPets('1')).toBe(await result);
    });
  });

  describe('fetchPets with an invalid id', () => {
    it('should throw an exception', async () => {
      expect(
        // Typecast invalid string as number
        petsService.getAll(('a' as unknown) as number),
      ).rejects.toThrow();
    });
  });

  describe('create pet should succeed with valid body', () => {
    it('should return the created pet', async () => {
      const result: Promise<Partial<Pet>> = new Promise(resolve =>
        resolve({ id: 1, name: 'Spot', type: 'DOG' }),
      );
      jest.spyOn(petsService, 'create').mockImplementation(() => result);
      const petDTO = new CreatePetDTO();
      petDTO.name = 'Spot';
      petDTO.type = 'DOG';

      const savedPet = await petsController.addPet(petDTO);
      expect(savedPet.name).toBe(petDTO.name);
      expect(savedPet.type).toBe(petDTO.type);
    });
  });

  describe('create pet should fail with invalid body', () => {
    it('should throw an exception', async () => {
      const result: Promise<Partial<Pet>> = new Promise(resolve =>
        resolve({ id: 1, name: 'Spot', type: 'DOG' }),
      );
      jest.spyOn(petsService, 'create').mockImplementation(() => result);
      const petDTO = new CreatePetDTO();
      petDTO.name = 'Spot';

      // Test failing due to @UsePipes(@ new ValidationPipe) not being injected
      // to the controller when tested. We will have to test validation logic in E2E tests.
      //expect(petsController.addPet(petDTO)).rejects.toThrow();
    });
  });
});
